move all monitor logic to dedicated module
